Refactored MapContext to MapContextForm.
