+++ frontend/src/Components/MapContextForm/MapContextForm.css
@@ -1,6 +1,10 @@
 .map-context {
-  width:calc(100% + 48px); /* to overlap parent's padding defined in site-layout class */
-  height:calc(100% + 48px); /* to overlap parent's padding defined in site-layout class */
+  width: calc(
+    100% + 48px
+  ); /* to overlap parent's padding defined in site-layout class */
+  height: calc(
+    100% + 48px
+  ); /* to overlap parent's padding defined in site-layout class */
   position: relative;
   margin-left: -24px; /* to overlap parent's padding defined in site-layout class */
   margin-top: -24px; /* to overlap parent's padding defined in site-layout class */
